# chore(handler): use optimized serializer to list applications (hl-1432)