## Description :sparkles:

DEV / TEST environment has begun to slow down in handler views where applications are listed in numbers. Currently, in TEST, requesting the API endpoint directly **to list only 26 applications takes around 2 - 3 seconds**. In frontend it takes even more time as there's multiple requests done simultaneously. Needless to say, this would also affect production at some point.

Here, a more simpler serializer is used. This is continued work and learnings from archive / search listings, where application count would've been a major bottle neck with the "include everything" `HandlerApplicationSerializer`. The results speak for themselves.

### Results

---

Listing 1700 applications in handler frontpage:

* Before: 40 - 60s
* After: 3 - 5s

---

A more "real world" case, listing 170 applications in handler frontpage:

* Before: 2 - 3s
* After: 0.3 - 0.5s

---

### Test

Tests can be repeated by checking in/out of this branch and using the following command in shell:

```
# Seeding

docker exec -it --user root benefit-backend python manage.py seed --number=100
# OR 
docker exec -it --user root benefit-backend python manage.py seed --number=10

# Timing
# Note: MacOS does not support "date +%N", thus node is used to get ms

node -e 'console.log(Date.now())' && curl https://localhost:8000/v1/handlerapplications/simplified_list/\?ahjo_case\=0\&format\=json\&order_by\=-submitted_at\&status\=received%2Chandling%2Cadditional_information_needed%2Caccepted%2Cdraft%2Crejected &> /dev/null &&  node -e 'console.log(Date.now())'
```