@terovirtanen 2024-08-28
**HANDLER branch is deployed to platta: https://helsinkilisa-ui-handler-pr3226.dev.hel.ninja** :rocket::rocket::rocket: