@terovirtanen 2024-08-28
**APPLICANT branch is deployed to platta: https://helsinkilisa-ui-pr3226.dev.hel.ninja** :rocket::rocket::rocket: