@terovirtanen 2024-08-28
## TestCafe result is __success__ for https://helsinkilisa-ui-pr3226.dev.hel.ninja :laughing::tada::tada::tada: