@terovirtanen 2024-08-28
## TestCafe handler result is __failed__ for https://helsinkilisa-ui-handler-pr3226.dev.hel.ninja :crying_cat_face::anger::boom::boom: