@terovirtanen 2024-08-29
**HANDLER branch is deployed to platta: https://helsinkilisa-ui-handler-pr3226.dev.hel.ninja** :rocket::rocket::rocket: