@terovirtanen 2024-08-29
**API branch is deployed to platta: https://helsinkilisa-pr3226.api.dev.hel.ninja/healthz** :rocket::rocket::rocket: